begin redoing api app endpoints
