+++ apps/api/logic.py
@@ -48,6 +48,28 @@ def add_new_user(db:str, resp_newuser:NewUser)->int:
         return user_id
 
 
+#Get user info given ID
+def get_user_info(id:int, db:str)->dict:
+    user_dict={}
+    try:
+        conn, cur = db_connect(db)
+        cur.execute("SELECT * FROM users WHERE user_id=%s",(id,))
+        uinfo = cur.fetchone()
+        # if no user_name maches given user_id
+        if not uinfo:
+            code = 404
+        else:  
+            user_dict['user_id'] = uinfo[0]
+            user_dict['user_name'] = uinfo[1]
+            user_dict['dob'] = uinfo[2]
+            user_dict['sex'] = uinfo[3]
+            user_dict['team'] = uinfo[4]
+            code = 200
+    finally:
+        cur.close()
+        conn.close()
+        return {'status_code': code, 'message':user_dict} 
+
 #Get user_name from ID
 def get_user_name(id, db):
     try:
