+++ ocr/pipeline.py
@@ -1,0 +1,63 @@
+import cv2
+import pytesseract
+import numpy as np
+
+## Image OCR
+def preprocess(image): #image
+    gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
+    bw_gaussian = cv2.adaptiveThreshold(gray,255,cv2.ADAPTIVE_THRESH_GAUSSIAN_C,cv2.THRESH_BINARY,11,6)
+    # noise removal
+    image = bw_gaussian
+    kernel = np.ones((2,2), np.uint8)
+    image = cv2.erode(image, kernel, iterations=1)
+    kernel = np.ones((2, 2), np.uint8)
+    image = cv2.dilate(image, kernel, iterations=1)
+    image = cv2.morphologyEx(image, cv2.MORPH_CLOSE, kernel)
+    image = cv2.medianBlur(image, 3)
+    image = cv2.cvtColor(image, cv2.COLOR_GRAY2RGB)
+    return image 
+
+def snip(img)->dict:
+    sr = img[182:400, 380:440]
+    split = img[180:400, 280:385]
+    dist = img[180:400, 180:280]
+    time = img[180:400, 50:180]
+    summary = img[135:175, 50:480]
+    date = img[47:100, 10:225]
+    workout = img[15:55, 10:150]
+    snips = {'wo':workout, 'date':date, 'summary':summary, 'time':time, 'dist':dist, 'split':split, 'sr':sr}
+    return snips
+
+def extract_data(snips:dict)->dict:
+    ocr_dict = {}
+    for s in snips:
+        if s in ['wo', 'date', 'summary']:
+            psm = '--psm 13'
+        else:
+            psm ='--psm 6'
+        gray = cv2.cvtColor(snips[s], cv2.COLOR_BGR2GRAY)
+        blur = cv2.GaussianBlur(gray, (1,1), 0)
+        thresh = cv2.threshold(blur, 0, 255, cv2.THRESH_BINARY_INV + cv2.THRESH_OTSU)[1]
+        data = pytesseract.image_to_string(thresh, lang='eng', config=psm)
+        ocr_dict[s] = data
+    return ocr_dict
+
+def basic_clean(ocr_dict:dict):
+    for key in ocr_dict:
+        ocr_dict[key] = ocr_dict[key].strip('\n\'"~|°`‘-!“ [()><')
+        ocr_dict[key]=ocr_dict[key].replace("\n", ' ')
+        ocr_dict[key]=ocr_dict[key].replace(",", '.')
+        s2 = ""
+        for i in range(len(ocr_dict[key])):
+            if not ocr_dict[key][i] in ['\\','|','-', '“']:
+                s2 += ocr_dict[key][i]
+        ocr_dict[key] = s2
+        if not key == 'date':
+            ocr_dict[key]=ocr_dict[key].replace("A", '4') 
+        if key != 'date':
+            ocr_dict[key] = ocr_dict[key].split()         
+    return ocr_dict
+
+#put ocr extraction into one func
+def clean_ocr(image):
+    return basic_clean(extract_data(snip(preprocess(image))))
